Grid Clean-Up Renaming card-grid to content-grid
