+++ kolibri/plugins/learn/assets/src/views/content-grid/content-grid-item.vue
