+++ kolibri/plugins/learn/assets/src/views/content-grid/index.vue
